Add helm station for ship controlls
The ship is commposed of two parts: internal and external. Internal is
everything the player can interact with, and includes players. External
is everything that interact with the outside world. When a player uses
the helm, they control the external, and the internal and players are
made to follow along with it. Following has a slight error due to floating
point precission, but it shouldnt be an issue.