[CODE][ADD/FIX][Add astroid event, Fix interactable bug]
The astroid event is one that can occur on the front, left, or right
region, and will cause damage to a random room in that region. It uses
the new hull manager to get a room belonging to a region.

If an interactable was destroied, the interactor that was within its
trigger would not be notified. This has been fixed.